对象池更新：placement new且类型系统上线 类型系统稳步更新 加强编译器功能

+++ src/include/code_loader.h
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-extern map<string, int> codes_int;
+extern map<string, short> codes_int;
 extern string int_code[];
 
 #endif
